+++ extensions/homeassistant/src/components/menu.tsx
@@ -9,9 +9,14 @@ import {
   Image,
   launchCommand,
   LaunchType,
+  environment,
 } from "@raycast/api";
 import { getErrorMessage } from "../utils";
 import { ReactNode } from "react";
+
+export async function relaunchCurrentCommand() {
+  await launchCommand({ name: environment.commandName, type: LaunchType.UserInitiated });
+}
 
 export function MenuBarItemConfigureCommand(): JSX.Element {
   return (
@@ -61,7 +66,7 @@ export function MenuBarSubmenu(props: {
   const sep = props.separator && props.separator.length > 0 ? props.separator : "|";
   const title =
     joinNonEmpty(
-      [props.title, sep, props.subtitle].filter((e) => e),
+      [props.title, props.subtitle && props.subtitle.length > 0 ? sep : undefined, props.subtitle].filter((e) => e),
       " ",
     ) || "";
   return (