+++ extensions/homeassistant/src/components/states/menu.tsx
@@ -6,6 +6,7 @@ import { CoverMenubarItem } from "../cover/menu";
 import { PersonMenubarItem } from "../persons/menu";
 import { SwitchMenubarItem } from "../switches/menu";
 import { CopyToClipboardMenubarItem, MenuBarSubmenu } from "../menu";
+import { LightMenubarItem } from "../lights/menu";
 
 export function CopyEntityIDToClipboard(props: { state: State }) {
   const s = props.state;
@@ -17,26 +18,21 @@ export function StateMenubarItem(props: { state: State }): JSX.Element | null {
   const e = s.entity_id;
   const domain = e.split(".")[0];
   switch (domain) {
-    case "media_player":
-      {
-        return <MediaPlayerMenubarItem state={s} />;
-      }
-      break;
-    case "cover":
-      {
-        return <CoverMenubarItem state={s} />;
-      }
-      break;
-    case "person":
-      {
-        return <PersonMenubarItem state={s} />;
-      }
-      break;
-    case "switch":
-      {
-        return <SwitchMenubarItem state={s} />;
-      }
-      break;
+    case "media_player": {
+      return <MediaPlayerMenubarItem state={s} />;
+    }
+    case "cover": {
+      return <CoverMenubarItem state={s} />;
+    }
+    case "person": {
+      return <PersonMenubarItem state={s} />;
+    }
+    case "switch": {
+      return <SwitchMenubarItem state={s} />;
+    }
+    case "light": {
+      return <LightMenubarItem state={s} />;
+    }
   }
   return (
     <MenuBarSubmenu key={s.entity_id} title={getFriendlyName(s)} subtitle={s.state} icon={getIcon(s)}>